refactor: gateway-service 패키지 구조 계층별로 리팩토링
