+++ gateway-service/src/main/java/com/carenest/business/gatewayservice/infrasturcture/config/GatewayConfig.java
@@ -1,5 +1,6 @@
-package com.carenest.business.gatewayservice;
+package com.carenest.business.gatewayservice.infrasturcture.config;
 
+import com.carenest.business.gatewayservice.infrasturcture.sercurity.JwtAuthFilter;
 import org.springframework.cloud.gateway.route.RouteLocator;
 import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
 import org.springframework.context.annotation.Bean;
