+++ gateway-service/src/main/java/com/carenest/business/gatewayservice/infrasturcture/sercurity/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.carenest.business.gatewayservice;
+package com.carenest.business.gatewayservice.infrasturcture.sercurity;
 
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
